Update tick.py
This will make the LEDs turn on in succession up to white, then turn off in reverse order and repeat.
To change the delay between the LED cycle, set sleep = [desired delay in s]
To change the brightness of the LEDs, set the value of the colour (piglow.green(#)) to desired brightness (0-255)
  
eg. piglow.green(1) ~ sets the brightness of the green LEDs to 1.